+++ src/routes/post_route.ts
@@ -2,56 +2,79 @@ import {Router, Response, Request} from "express";
 import {RequestWithBody, RequestWithParams, RequestWithParamsAndBody} from "../types/common";
 import {authMiddleware} from "../middlewares/auth/auth_middleware";
 import {CreatePostModel, UpdatePostModel} from "../types/posts/input";
-import {PostRepository} from "../repositories/post_repository";
+import {PostRepository} from "../repositories/post_db_repository";
 import {postValidation} from "../validators/post-validator";
+import {ObjectId} from "mongodb";
+import {BlogRepository} from "../repositories/blog_db_repository";
 
 export const postRoute = Router({})
 
-postRoute.get('/', (req: Request, res:Response) => {
-    const posts = PostRepository.getAllPosts()
+postRoute.get('/', async (req: Request, res:Response) => {
+    const posts = await PostRepository.getAllPosts()
 
-    res.send(posts)
+    res.status(200).send(posts)
 })
 
-postRoute.get('/:id', (req: RequestWithParams<{id: string}>, res:Response) => {
+postRoute.get('/:id', async (req: RequestWithParams<{id: string}>, res:Response) => {
     const id = req.params.id
-    const post = PostRepository.getPostById(id)
+    if (!ObjectId.isValid(id)){
+        res.sendStatus(404)
+        return
+    }
+
+    const post = await PostRepository.getPostById(id)
 
     if (!post) {
         res.sendStatus(404)
+        return
     }
 
-    res.send(post)
+    res.status(200).send(post)
 })
 
-postRoute.post('/', authMiddleware, postValidation(), (req: RequestWithBody<CreatePostModel>, res:Response) => {
+postRoute.post('/', authMiddleware, postValidation(), async (req: RequestWithBody<CreatePostModel>, res:Response) => {
     let {title, shortDescription, content, blogId} = req.body
-    const newPost = PostRepository.createPost(title, shortDescription, content, blogId)
+    const newPost = await PostRepository.createPost({title, shortDescription, content, blogId})
 
-    if (newPost) {
-        res.status(201).send(newPost)
+    const checkInsertion = await PostRepository.getPostById(newPost.id)
+
+    if (!checkInsertion) {
+        res.sendStatus(400)
     }
+    res.status(201).send(checkInsertion)
 })
 
-postRoute.put('/:id', authMiddleware, postValidation(), (req: RequestWithParamsAndBody<{id: string},UpdatePostModel>, res:Response) => {
+postRoute.put('/:id', authMiddleware, postValidation(), async (req: RequestWithParamsAndBody<{id: string},UpdatePostModel>, res:Response) => {
     const id = req.params.id
+    if (!ObjectId.isValid(id)){
+        res.sendStatus(404)
+        return
+    }
+
     let {title, shortDescription, content, blogId} = req.body
-    const updatedPostFlag = PostRepository.updatePostById(id, title, shortDescription, content, blogId)
+    const updatedPostFlag = await PostRepository.updatePostById(id, {title, shortDescription, content, blogId})
 
-    if (updatedPostFlag) {
-        res.sendStatus(204)
-    } else {
+    if (!updatedPostFlag) {
         res.sendStatus(404)
+        return
     }
+
+    res.sendStatus(204)
+
 })
 
-postRoute.delete('/:id', authMiddleware,(req: RequestWithParams<{id: string}>, res:Response) => {
+postRoute.delete('/:id', authMiddleware,async (req: RequestWithParams<{id: string}>, res:Response) => {
     const id = req.params.id
-    const deletedPostFlag = PostRepository.deletePostById(id)
-
-    if (deletedPostFlag) {
-        res.sendStatus(204)
-    } else {
+    if (!ObjectId.isValid(id)){
         res.sendStatus(404)
+        return
     }
+
+    const deletedPostFlag = await PostRepository.deletePostById(id)
+
+    if (!deletedPostFlag) {
+        res.sendStatus(404)
+        return
+    }
+    res.sendStatus(204)
 })
